migrate p-bus to "memory" bus to avoid usage conflicts
If we leave the p-bus in peripherals, LiteX will try to use it
for its own puproses. For the core, we want the bus to be
"pristine", so, we break it out as a notional "memory" bus
even though eventually we'll use it for peripherals.

Note that the metadata says the bus is cached, but in reality,
the core gen is not cached.