add explicit `.re` for axi-lite busses
The current CSR implementation infers the "read enable" signal
(which is, oddly enough, called `.we` in the code) by taking
the inverse of the write enable and combining it with the current
address on the bus.

This causes a problem for AXI-lite per issue #1530, where the read
enable signal can be asserted for several cycles instead of exactly
one per the CSR specification.

I tried addressing this in PR#1531 but the patch isn't good enough;
especially with third-party (non-Litex) IP, it is legal
to have the address be present on the bus for much longer than
one cycle, which leads to the read enable signal being asserted
for longer than one cycle.

This commit adds an explicit `.re` signal to the bus, so that this
can be derived from the "ready" signal on the AXI-lite bus, allowing
us to have a robust solution to guaranteeing the CSR abstraction is
met.